+++ src/test/java/org/springframework/test/ioc/CircularReferenceTest.java
@@ -33,4 +33,15 @@ public class CircularReferenceTest {
     // a被代理，放进二级缓存earlySingletonObjects中的是实例化后的A，而放进一级缓存singletonObjects中的是被代理后的A，实例化b时从earlySingletonObjects获取a，所以b.getA() != a
     assertThat(b.getA() != a).isTrue();
   }
+
+  @Test
+  public void test_circularReferenceWithProxy() throws Exception {
+    ClassPathXmlApplicationContext applicationContext =
+        new ClassPathXmlApplicationContext("classpath:spring-circular-reference-3.xml");
+    A a = applicationContext.getBean("a", A.class);
+    B b = applicationContext.getBean("b", B.class);
+
+    // 增加三级缓存不能解决有代理对象时的循环依赖。
+    assertThat(b.getA() == a).isTrue();
+  }
 }